Fix conflicts with localStorage when updating tasks status
